requirements.txt
sortedcontainers==1.5.7
regex==2017.7.28
lxml==3.8.0
idna==2.5
certifi==2017.1.23
cffi==1.9.1
requests[security]==2.13.0
Kivy-Garden==0.1.4
# Setting this because the wheel has been broken before and we'd rather build it ourselves
git+https://github.com/kivy/kivy.git@2.0.0rc3
PyInstaller==3.5
Pillow==4.1.1
PyYAML==3.10
MarkupSafe
Mako==0.5.0
args==0.1.0
clint==0.3.3
schema==0.5.0
docopt==0.6.1
internetarchive
jsonpatch>=0.4
pytest==2.3.4
futures==3.0.2
pycparser==2.17
pexpect==3.1
plyer
py==1.4.23
pyasn1==0.1.9
cryptography==2.2.1
pyOpenSSL==16.2.0
pycrypto==2.6.1
qrcode==5.3
raven==5.7.2
pycups==1.9.70
rstr==2.2.3
xml2json==1.1
isbnlib==3.8.3
psutil==5.4.7
unidecode==1.0.22
#timeago==1.0.8
prompt-toolkit==2.0.7
wcwidth==0.1.8
fysom==2.1.5
networkx==2.2
pyserial==3.4
Twisted==19.7.0
Werkzeug==0.16.0
Flask==1.1.1
m2r==0.2.1
coverage


# Please note that python-apt is best installed via APT
#python-apt==0.9.3.5ubuntu1
